Dialogs fixes and enhancements: - correct validation for numbers in input dialog - show markers for invalid values in dialog - remove standard header text from prompt

+++ utility-fx-controls/src/main/java/com/dua3/utility/fx/controls/InputDialogBuilder.java
@@ -139,7 +139,7 @@ public class InputDialogBuilder
     }
 
     @Override
-    public InputDialogBuilder integer(String id, String label, Supplier<Integer> dflt, Function<Integer, Optional<String>> validate) {
+    public InputDialogBuilder integer(String id, String label, Supplier<Long> dflt, Function<Long, Optional<String>> validate) {
         pb.integer(id, label, dflt, validate);
         return this;
     }
